\name{HairColor}
\alias{HairColor}
\docType{data}
\title{
Demographic Data
}
\description{
The sample data corresponds to 7 people and shows some demographic
  characteristics.
}
\usage{data(HairColor)}
\format{
  This data frame contains 7 observation for the following 5 columns:
  \describe{
    \item{Sex}{two levels (M=male,F=female)}  
    \item{HairColor}{four levels of hair color (Dark, Grey, Fair and Brown)}
    \item{Region}{(E = England,S = Scotland, W = Wales)}
    \item{Work}{(Manual,Clerical,Professional)}
    \item{Education}{(School,Univ,Postgrad)}
  }
}
\source{
Gower, J., Gardner-Lubbe,S., Le Roux,N. (2011). \dQuote{Understanding Biplots.} \emph{Wiley}.
}
\examples{
data(HairColor)
}
\keyword{datasets}
